Fix error in Vacantion

+++ Vacantion/Program.cs
@@ -10,27 +10,27 @@ namespace Vacantion
     {
         static void Main(string[] args)
         {
-            var money = double.Parse(Console.ReadLine());
+            var money = decimal.Parse(Console.ReadLine());
             var season = Console.ReadLine().ToLower();
             if (season == "winter")
             {
                 if (money<=1000)
                 {
-                    double pay = money * 45 / 100;
+                    decimal pay = money * 45 / 100;
                     Console.WriteLine("Morocco - Camp - {0:0.00}", Math.Round(pay,2));
                 }
                 else
                 {
                     if (money > 1000 && money <= 3000)
                     {
-                        double pay = money * 60 / 100;
+                        decimal pay = money * 60 / 100;
                         Console.WriteLine("Morocco - Hut - {0:0.00}", Math.Round(pay, 2));
                     }
                     else
                     {
                         if (money > 3000)
                         {
-                            double pay = money * 90 / 100;
+                            decimal pay = money * 90 / 100;
                             Console.WriteLine("Morocco - Hotel - {0:0.00}", Math.Round(pay, 2));
                         }
                     }
@@ -40,21 +40,21 @@ namespace Vacantion
             {
                 if (money <= 1000)
                 {
-                   double pay = money * 65 / 100;
+                    decimal pay = money * 65 / 100;
                     Console.WriteLine("Alaska - Camp - {0:0.00}", Math.Round(pay, 2));
                 }
                 else
                 {
                     if (money > 1000 && money <= 3000)
                     {
-                       double pay = money * 80 / 100;
+                        decimal pay = money * 80 / 100;
                         Console.WriteLine("Alaska - Hut - {0:0.00}", Math.Round(pay, 2));
                     }
                     else
                     {
                         if (money > 3000)
                         {
-                            double pay = money * 90 / 100;
+                            decimal pay = money * 90 / 100;
                             Console.WriteLine("Alaska - Hotel - {0:0.00}", Math.Round(pay, 2));
                         }
                     }
